hello_world/requirements.txt
flask-lambda-python36==0.1.0
requests==2.22.0

